gremlin/chatbot-full-stack-application/deployment/requirements.txt
amazon-lex-bot-deploy==0.1.9
beautifulsoup4==4.9.3
boto3==1.16.4
botocore==1.19.4
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
jmespath==0.10.0
Pillow==8.0.1
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
tenacity==6.2.0
urllib3==1.25.11
